+++ src/include/EvCheckBox.h
@@ -24,7 +24,7 @@ class EvCheckBox : public EvObj
     bool          SetValue(int16_t Value);
     void          SetColor(uint16_t ColorCheck, uint16_t ColorUncheck, uint16_t ColorBorder);
     void          SetOnTouch(void (*OnTouch)(EvCheckBox *Sender, EvTouchEvent *Touch));
-    void          SetOnChange(void (*OnChange)(EvCheckBox *Sender, int16_t Value));
+    void          SetOnChange(void (*OnChange)(EvCheckBox *Sender, int32_t Value));
 
   protected:
     EvCheckBox(int16_t Left, int16_t Top, uint16_t Width, uint16_t Height, EvDisplay *Disp, const char *Tag = NULL, uint16_t State = VISIBLE_OBJ);
@@ -40,7 +40,7 @@ class EvCheckBox : public EvObj
 
   private:
     void          (*mOnTouch)(EvCheckBox *Sender, EvTouchEvent *Touch);
-    void          (*mOnChange)(EvCheckBox *Sender, int16_t Value);
+    void          (*mOnChange)(EvCheckBox *Sender, int32_t Value);
 
   public:
     static EvCheckBox *Create(int16_t Left, int16_t Top, uint16_t Width, uint16_t Height, EvPanel *Dest, const char *Tag = NULL, uint16_t State = VISIBLE_OBJ);
